requirements.txt
fastapi
uvicorn[standard]
google-ads
pydantic
pandas
pyyaml
python-multipart
python-dotenv
